MNT: switch to pip-installable lxml for best compatibility
