+++ forTwinCatRepos/.pre-commit-config.yaml
@@ -11,4 +11,4 @@ repos:
     hooks:
     -   id: twincat-leading-tabs-remover
     -   id: twincat-lineids-remover
-    -   id: twincat-xmllint-format
+    -   id: twincat-xml-format
