+++ requirements.txt
@@ -1,1 +1,1 @@
-libxml2-python3
+lxml
